boop

+++ src/felix86/hle/syscall.cpp
@@ -798,7 +798,7 @@ void felix86_syscall(ThreadState* state) {
             }
         }
         FELIX86_UNLOCK;
-        if (r8 != -1) { // Uses file descriptor for mmap, may need to update symbols
+        if ((int)r8 != -1) { // Uses file descriptor for mmap, may need to update symbols
             g_cached_symbols = false;
         }
         break;
